pyproject.toml
[project]
name = "crewai-gmail-cleaner"
version = "0.1.0"
description = "CrewAI crew to clean Gmail (Plain/Linear notifications)"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "crewai",
    "google-api-python-client",
    "google-auth",
    "google-auth-oauthlib",
]

[tool.crewai]
type = "crew"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/crewai_gmail_cleaner"]
